Statistical-Analysis/Hypothesis-Testing-College-Data/requirements.txt
pandas
numpy
scipy.stats